requirements.txt
click==8.0.4
eyed3==0.9.6
ffmpeg-python==0.2.0
pandas==1.4.1
rich==11.2.0
scandir==1.10.0
tinytag==1.7.0
